refactor: split App into separate components (Ball, Walls, Floor, Bumper, Flipper)
